pyproject.toml
[project]
name = "cc-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "python-frontmatter>=1.1.0",
]
